trying deleting user

+++ server/routes/Users.js
@@ -19,13 +19,18 @@ router.post("/", async (req, res) => {
 //delete user
 router.delete("/deleteuser/:id", validateToken, async (req, res) => {
   const userId = req.params.id;
-  await Users.destroy({
-    where: {
-      id: userId,
-    },
-  });
-  res.json(userId);
-  //res.json(`USER NBR ${userId} DELETED SUCCESSFULLY`);
+  const userExist = await Users.findOne({ where: { id: userId } });
+
+  if (!userExist) {
+    res.json({ error: "User Doesn't Exist" });
+  } else {
+    await Users.destroy({
+      where: {
+        id: userId,
+      },
+    });
+    res.json(`USER NBR ${userId} DELETED SUCCESSFULLY`);
+  }
 });
 
 router.post("/login", async (req, res) => {
